Adjusted Export to Excel function
Some variables from bup_plan_analyzer became global in order to access it from BUP_GUI main function, as bup.variable. This is because get_entry_values() does not return anything, as it is called by a command argument from a CTkButton. Now, when user clicks in export button, an Excel file with 2 tabs are exported. Fully functional.